policy ok admon en cours

+++ app/Http/Controllers/SkillController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers;
 use App\SkillUser;
 
 //use DebugBar\DebugBar;
-use DebugBar\DebugBar;
+use DebugBar;
 use Illuminate\Http\Request;
 use App\Skill;
 use App\User;
@@ -55,19 +55,23 @@ class SkillController extends Controller
      */
     public function store(Request $request, User $user)
     {
+        //\DebugBar::message($user);
+//        \debugbar()->addMessage($user);
 //        $this->authorize('store',$user->getEmail());
 //        $this->authorize('store', $user);
-        $request->validate([
-            'Skill'=>'required',
-        ]);
-
-        $Us = User::find($request->post("id"));
-        $skill= Skill::find($request->post("Skill"));
-        $Us->skills()->attach($skill, array('level' => 0));
-        $id=$request->post("id");
-        $redi ="/skill/$id/edit";
-        return redirect($redi);
-
+        if (auth()->user()->can('store', Skill::class)) {
+            $request->validate([
+                'Skill' => 'required',
+            ]);
+//        dd($request->post("id"));
+            $Us = User::find($request->post("id"));
+            $skill = Skill::find($request->post("Skill"));
+            $Us->skills()->attach($skill, array('level' => 0));
+            $id = $request->post("id");
+            $redi = "/skill/$id/edit";
+            return redirect($redi);
+        }
+        return view('/home');
     }
 
     /**
@@ -88,16 +92,20 @@ class SkillController extends Controller
      * @return \Illuminate\Http\Response
      * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function edit($id,User $user)
+    public function edit($id)
     {
         //
-        $this->authorize('edit',$user);
-        $userSkill = User::find($id);
-        $skillsId=[];
-        foreach ($userSkill->skills()->get() as $element){
-            array_push($skillsId,$element->id);
+//        $this->authorize('edit',$user);
+        if(auth()->user()->can('edit', Skill::class)){
+            $userSkill = User::find($id);
+            $skillsId=[];
+            foreach ($userSkill->skills()->get() as $element){
+                array_push($skillsId,$element->id);
+            }
+            return view('skill.edit',compact('skillsId'));
+
         }
-        return view('skill.edit',compact('skillsId'));
+        return view('/home');
     }
 
     /**
@@ -111,16 +119,20 @@ class SkillController extends Controller
      */
     public function update(Request $request, $id,User $user)
     {
-        $this->authorize('update',$user);
-       foreach ($request->except('_token') as $idSkill =>$value){
-           $user = User::find($id);
-           $skill= Skill::find($idSkill);
 
-           $user->skills()->updateExistingPivot($skill, array('level' => $value), false);
-       }
-//        $redi ="/skill/$id/edit";
+//        $this->authorize('update',$user);
+        if (auth()->user()->can('store', Skill::class)) {
+            foreach ($request->except('_token') as $idSkill => $value) {
+                $user = User::find($id);
+                $skill = Skill::find($idSkill);
+
+                $user->skills()->updateExistingPivot($skill, array('level' => $value), false);
+            }
+//            $redi = "/skill/$id/edit";
         return redirect('/home');
-       // return redirect('/home');
+//            return redirect($redi);
+        }
+        return redirect('/home');
     }
 
     /**
@@ -135,13 +147,15 @@ class SkillController extends Controller
     public function destroy(User $users, Request $request,$id)
     {
         //
-//        $this->authorize('delete',$users);
-        $request->validate([
-            'Skill'=>'required',
-        ]);
-        $user = User::find($id);
-        $user->skills()->detach($request->post("Skill"));
-        $redi ="/skill/$id/edit";
-        return redirect($redi);
+        if (auth()->user()->can('edit', Skill::class)) {
+            $request->validate([
+                'Skill' => 'required',
+            ]);
+            $user = User::find($id);
+            $user->skills()->detach($request->post("Skill"));
+            $redi = "/skill/$id/edit";
+            return redirect($redi);
+        }
+        return view('/home');
     }
 }
